+++ packages/backend/src/modules/tracked-txs/TrackedTxsModule.ts
@@ -19,7 +19,6 @@ import { L2CostsPricesIndexer } from './modules/l2-costs/indexers/L2CostsPricesI
 import { createLivenessModule } from './modules/liveness/LivenessModule'
 import { AnomaliesIndexer } from './modules/liveness/indexers/AnomaliesIndexer'
 import { LivenessAggregatingIndexer } from './modules/liveness/indexers/LivenessAggregatingIndexer'
-import { LivenessAggregatingIndexer2 } from './modules/liveness/indexers/LivenessAggregatingIndexer2'
 
 export function createTrackedTxsModule(
   config: Config,
@@ -112,20 +111,10 @@ export function createTrackedTxsModule(
   }
 
   let livenessAggregatingIndexer: LivenessAggregatingIndexer | undefined
-  let livenessAggregatingIndexer2: LivenessAggregatingIndexer2 | undefined
   let anomaliesIndexer: AnomaliesIndexer | undefined
 
   if (config.trackedTxsConfig.uses.liveness) {
     livenessAggregatingIndexer = new LivenessAggregatingIndexer({
-      db: peripherals.database,
-      projects: config.trackedTxsConfig.projects,
-      parents: [trackedTxsIndexer],
-      indexerService,
-      minHeight: config.trackedTxsConfig.minTimestamp,
-      logger: logger.tag({ feature: 'liveness' }),
-    })
-
-    livenessAggregatingIndexer2 = new LivenessAggregatingIndexer2({
       db: peripherals.database,
       projects: config.trackedTxsConfig.projects,
       parents: [trackedTxsIndexer],
@@ -156,7 +145,6 @@ export function createTrackedTxsModule(
     await l2CostPricesIndexer?.start()
     await l2CostsAggregatorIndexer?.start()
     await livenessAggregatingIndexer?.start()
-    await livenessAggregatingIndexer2?.start()
     await anomaliesIndexer?.start()
   }
